[update] ESLintでエラーが出てたので諸々変更
ESLint: Maximum call stack size exceeded Occurred while linting /home/runner/work/homimemo/homimemo/src/app/400/page.tsx:1 Rule: "@stylistic/indent"
