+++ package.json
@@ -58,9 +58,13 @@
   },
   "devDependencies": {
     "@chromatic-com/storybook": "^4.1.1",
+    "@eslint/css": "^0.11.0",
     "@eslint/eslintrc": "^3.3.1",
+    "@eslint/js": "^9.35.0",
+    "@eslint/json": "^0.13.2",
     "@markuplint/jsx-parser": "^4.7.20",
     "@markuplint/react-spec": "^4.5.20",
+    "@next/eslint-plugin-next": "^15.5.2",
     "@storybook/addon-docs": "^9.1.3",
     "@storybook/addon-onboarding": "^9.1.3",
     "@storybook/addon-vitest": "^9.1.3",
@@ -75,20 +79,25 @@
     "@vitest/browser": "^3.2.4",
     "@vitest/coverage-v8": "^3.2.4",
     "chromatic": "^11.29.0",
-    "eslint": "^9.33.0",
+    "eslint": "^9.35.0",
     "eslint-config-next": "15.3.0",
     "eslint-plugin-jsdoc": "^50.8.0",
-    "eslint-plugin-storybook": "^9.1.3",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "eslint-plugin-storybook": "^9.1.5",
     "feed": "^5.1.0",
+    "globals": "^16.3.0",
+    "jiti": "^2.5.1",
     "markuplint": "^4.13.0",
     "npm-run-all": "^4.1.5",
     "playwright": "^1.55.0",
     "storybook": "^9.1.3",
-    "stylelint": "^16.23.1",
+    "stylelint": "^16.24.0",
     "stylelint-config-standard": "^38.0.0",
     "svgo": "^4.0.0",
     "tsx": "^4.20.4",
     "typescript": "^5.9.2",
+    "typescript-eslint": "^8.42.0",
     "vite-plugin-svgr": "^4.3.0",
     "vitest": "^3.2.4",
     "wrangler": "^4.33.2"
